Refactor fabrid accumulator
