+++ daemon/internal/servers/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//pkg/addr:go_default_library",
         "//pkg/drkey:go_default_library",
         "//pkg/experimental/fabrid:go_default_library",
+        "//pkg/experimental/fabrid/graphutils:go_default_library",
         "//pkg/grpc:go_default_library",
         "//pkg/log:go_default_library",
         "//pkg/metrics:go_default_library",
@@ -28,7 +29,6 @@ go_library(
         "//pkg/segment/extensions/fabrid:go_default_library",
         "//pkg/snet:go_default_library",
         "//pkg/snet/path:go_default_library",
-        "//private/path/combinator:go_default_library",
         "//private/revcache:go_default_library",
         "//private/topology:go_default_library",
         "//private/trust:go_default_library",
