+++ pkg/experimental/fabrid/graphutils/BUILD.bazel
@@ -1,0 +1,14 @@
+load("//tools/lint:go.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["maps.go"],
+    importpath = "github.com/scionproto/scion/pkg/experimental/fabrid/graphutils",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/addr:go_default_library",
+        "//pkg/experimental/fabrid:go_default_library",
+        "//pkg/segment/extensions/fabrid:go_default_library",
+        "//pkg/snet:go_default_library",
+    ],
+)
