+++ private/path/combinator/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "combinator.go",
-        "fabrid_accumulator.go",
         "graph.go",
         "staticinfo_accumulator.go",
     ],
@@ -12,12 +11,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/addr:go_default_library",
-        "//pkg/experimental/fabrid:go_default_library",
+        "//pkg/experimental/fabrid/graphutils:go_default_library",
         "//pkg/private/common:go_default_library",
         "//pkg/private/ctrl/path_mgmt/proto:go_default_library",
         "//pkg/private/util:go_default_library",
         "//pkg/segment:go_default_library",
-        "//pkg/segment/extensions/fabrid:go_default_library",
         "//pkg/segment/extensions/staticinfo:go_default_library",
         "//pkg/slayers/path:go_default_library",
         "//pkg/slayers/path/scion:go_default_library",
