+++ private/path/combinator/graph.go
@@ -23,6 +23,7 @@ import (
 	"time"
 
 	"github.com/scionproto/scion/pkg/addr"
+	fabrid_utils "github.com/scionproto/scion/pkg/experimental/fabrid/graphutils"
 	"github.com/scionproto/scion/pkg/private/common"
 	"github.com/scionproto/scion/pkg/private/ctrl/path_mgmt/proto"
 	"github.com/scionproto/scion/pkg/private/util"
@@ -314,7 +315,7 @@ type pathSolution struct {
 func (solution *pathSolution) Path() Path {
 	mtu := ^uint16(0)
 	var segments segmentList
-	fabridMaps := make(map[addr.IA]FabridMapEntry)
+	fabridMaps := make(map[addr.IA]fabrid_utils.FabridMapEntry)
 	var epicPathAuths [][]byte
 	for _, solEdge := range solution.edges {
 		var hops []path.HopField
@@ -384,7 +385,7 @@ func (solution *pathSolution) Path() Path {
 			fabridMap, exists := fabridMaps[asEntry.Local]
 			if (!exists || fabridMap.Ts.Before(solEdge.segment.Info.Timestamp)) && asEntry.
 				Extensions.Digests != nil {
-				fabridMaps[asEntry.Local] = FabridMapEntry{
+				fabridMaps[asEntry.Local] = fabrid_utils.FabridMapEntry{
 					Map:    asEntry.UnsignedExtensions.FabridDetached,
 					Digest: asEntry.Extensions.Digests.Fabrid.Digest,
 					Ts:     solEdge.segment.Info.Timestamp,
@@ -425,7 +426,7 @@ func (solution *pathSolution) Path() Path {
 	interfaces := segments.Interfaces()
 	asEntries := segments.ASEntries()
 	staticInfo := collectMetadata(interfaces, asEntries)
-	fabridInfo := collectFabridPolicies(interfaces, fabridMaps)
+	fabridInfo := fabrid_utils.CollectFabridPolicies(interfaces, fabridMaps)
 	path := Path{
 		SCIONPath: segments.ScionPath(),
 		Metadata: snet.PathMetadata{
